# docs/backend-contracts/contract-history/1.3.0-contract-914084874afda2481813a041da4cc01249aa9ea557d9a8bf29baeed4f10e0dc9/openapi.snapshot.yaml
openapi: 3.1.0
info:
  title: BNBU Sports Unified API
  version: 1.3.0-contract
  summary: Unified contract for the sports check-in platform
  description: |
    Stage 21B contract. It does not prove that every registered capability is
    business-enabled or production-ready.
    Business fields come from 02-data-dictionary.md, state transitions from
    03-state-machines.md, rules from 04-business-rules.md, and authorization
    from 05-permission-matrix.md. Foundation authentication/session persistence
    is approved; CourseInvite/Join Capability are contract-and-persistence-design
    only until the Enrollment Gate, while every Export operation is fail-closed
    and V1 persists no ExportJob. Of the 30 client-support operations added in
    Stage 21, authentication, exemption, release-policy, notification,
    push-device, preference, help, and feedback capabilities are approved for
    real local implementation, while unresolved capabilities remain explicit
    routed denials. GPS has persistence and
    application-layer foundations, but all six location HTTP operations remain
    denied. None of these statements proves Staging, iOS-binary, push-delivery,
    GPS-production, or Production-Gate readiness.
servers:
  - url: /api/v1
    description: Version 1 relative base path
security:
  - bearerAuth: []
...
paths:
...
  /enrollments:
    get:
      tags: [Enrollments]
      operationId: listEnrollments
      x-access-policy:
        policyId: ENROLLMENT-LIST
        authentication: ACCESS_TOKEN
        allowedRoles: [STUDENT, TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: ROLE_SCOPED
        resourceResolver: ENROLLMENT_LIST_SCOPE
        defaultDeny: true
      summary: List authorized enrollment projections
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/ClassSectionFilter'
        - name: studentId
          in: query
          schema: { $ref: '#/components/schemas/OpaqueId' }
        - name: semesterId
          in: query
          schema: { $ref: '#/components/schemas/OpaqueId' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/EnrollmentStatus' }
      responses:
        '200':
          $ref: '#/components/responses/EnrollmentListSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /class-sections/{classSectionId}/enrollments:
    post:
      tags: [Enrollments]
      operationId: manuallyEnrollStudent
      x-access-policy:
        policyId: ENROLLMENT-MANUAL-ADD
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: TEACHER_CLASS_SECTION
        resourceResolver: CLASS_SECTION_FROM_PATH
        defaultDeny: true
      summary: Manually add an existing student to a class section owned by the responsible teacher
      parameters:
        - $ref: '#/components/parameters/ClassSectionId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualEnrollmentRequest'
      responses:
        '201':
          $ref: '#/components/responses/EnrollmentSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /enrollments/{enrollmentId}:
    get:
      tags: [Enrollments]
      operationId: getEnrollment
      x-access-policy:
        policyId: ENROLLMENT-READ
        authentication: ACCESS_TOKEN
        allowedRoles: [STUDENT, TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: ROLE_SCOPED
        resourceResolver: ENROLLMENT_FROM_PATH
        defaultDeny: true
      summary: Get an authorized enrollment
      parameters:
        - $ref: '#/components/parameters/EnrollmentId'
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          $ref: '#/components/responses/EnrollmentSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /enrollments/{enrollmentId}/withdraw:
    post:
      tags: [Enrollments]
      operationId: withdrawEnrollment
      x-access-policy:
        policyId: ENROLLMENT-WITHDRAW
        authentication: ACCESS_TOKEN
        allowedRoles: [STUDENT]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: SELF
        resourceResolver: ENROLLMENT_FROM_PATH
        defaultDeny: true
      summary: Withdraw the authenticated student's active enrollment
      description: Contract placeholder. Always returns ENROLLMENT_WITHDRAWAL_DISABLED with no side effects until ADR-054 is accepted.
      x-enabled-by-default: false
      x-business-blocker: ADR-054
      x-default-deny-error: ENROLLMENT_WITHDRAWAL_DISABLED
      parameters:
        - $ref: '#/components/parameters/EnrollmentId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedReasonRequest'
      responses:
        '200':
          $ref: '#/components/responses/EnrollmentSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /enrollments/{enrollmentId}/remove:
    post:
      tags: [Enrollments]
      operationId: removeEnrollment
      x-access-policy:
        policyId: ENROLLMENT-REMOVE
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: TEACHER_CLASS_SECTION
        resourceResolver: ENROLLMENT_FROM_PATH
        defaultDeny: true
      summary: Remove a student from a class section owned by the responsible teacher
      parameters:
        - $ref: '#/components/parameters/EnrollmentId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedReasonRequest'
      responses:
        '200':
          $ref: '#/components/responses/EnrollmentSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /enrollments/{enrollmentId}/restore:
    post:
      tags: [Enrollments]
      operationId: restoreEnrollment
      x-access-policy:
        policyId: ENROLLMENT-RESTORE
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: TEACHER_CLASS_SECTION
        resourceResolver: ENROLLMENT_FROM_PATH
        defaultDeny: true
      summary: Restore a withdrawn or removed enrollment
      parameters:
        - $ref: '#/components/parameters/EnrollmentId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedReasonRequest'
      responses:
        '200':
          $ref: '#/components/responses/EnrollmentSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /course-invites/{inviteToken}/join:
    post:
      tags: [QR Course Joining, Enrollments]
      operationId: joinClassSectionWithInvite
      x-access-policy:
        policyId: ENROLLMENT-JOIN
        authentication: JOIN_CAPABILITY
        allowedRoles: []
        organizationScope: CAPABILITY_ORGANIZATION
        resourceScope: CAPABILITY_CLASS_SECTION
        resourceResolver: JOIN_CAPABILITY
        defaultDeny: true
      summary: Consume a Join Capability and atomically establish the student session
      description: No ordinary Access Token is accepted. One transaction consumes the capability and creates or reuses User, StudentProfile, ACTIVE Enrollment, and AuthSession; any failure rolls back all effects.
      security:
        - JoinCapability: []
      parameters:
        - $ref: '#/components/parameters/InviteToken'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      responses:
        '200':
          $ref: '#/components/responses/JoinSuccess'
        '201':
          $ref: '#/components/responses/JoinSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /exports:
    get:
      tags: [Exports]
      operationId: listExports
      x-access-policy:
        policyId: EXPORT-LIST
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: ROLE_SCOPED
        resourceResolver: EXPORT_LIST_SCOPE
        defaultDeny: true
      summary: List export jobs visible to the requesting actor
      description: Contract-only endpoint; V1 returns SYSTEM_MODE_UNSUPPORTED because no ExportJob is persisted.
      x-enabled-by-default: false
      x-default-deny-error: SYSTEM_MODE_UNSUPPORTED
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - name: exportType
          in: query
          schema: { $ref: '#/components/schemas/ExportType' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ExportStatus' }
      responses:
        '200':
          $ref: '#/components/responses/ExportListSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    post:
      tags: [Exports]
      operationId: createExport
      x-access-policy:
        policyId: EXPORT-CREATE
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: ROLE_SCOPED
        resourceResolver: EXPORT_SCOPE_FROM_BODY
        defaultDeny: true
      summary: Request an authorized asynchronous export
      description: Contract-only endpoint; V1 returns SYSTEM_MODE_UNSUPPORTED without creating an ExportJob or artifact.
      x-enabled-by-default: false
      x-default-deny-error: SYSTEM_MODE_UNSUPPORTED
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExportRequest'
      responses:
        '202':
          $ref: '#/components/responses/ExportSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /exports/{exportId}:
    get:
      tags: [Exports]
      operationId: getExport
      x-access-policy:
        policyId: EXPORT-READ
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: ROLE_SCOPED
        resourceResolver: EXPORT_FROM_PATH
        defaultDeny: true
      summary: Get an authorized export job status
      description: Contract-only endpoint; V1 returns SYSTEM_MODE_UNSUPPORTED because no ExportJob is persisted.
      x-enabled-by-default: false
      x-default-deny-error: SYSTEM_MODE_UNSUPPORTED
      parameters:
        - $ref: '#/components/parameters/ExportId'
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          $ref: '#/components/responses/ExportSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /exports/{exportId}/download-url:
    post:
      tags: [Exports]
      operationId: createExportDownloadUrl
      x-access-policy:
        policyId: EXPORT-DOWNLOAD-URL
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: ROLE_SCOPED
        resourceResolver: EXPORT_FROM_PATH
        defaultDeny: true
      summary: Create a short-lived download URL after rechecking export scope
      description: Contract-only endpoint; V1 returns SYSTEM_MODE_UNSUPPORTED without creating a signed URL.
      x-enabled-by-default: false
      x-default-deny-error: SYSTEM_MODE_UNSUPPORTED
      parameters:
        - $ref: '#/components/parameters/ExportId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                purpose:
                  type: string
                  minLength: 1
                  maxLength: 500
              required: [purpose]
      responses:
        '200':
          $ref: '#/components/responses/ExportDownloadSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token carried only in the Authorization header.
    JoinCapability:
      type: apiKey
      in: header
      name: X-Join-Capability
      description: Short-lived one-time capability bound to one ClassSection. The plaintext value is never logged or persisted; no ordinary Access Token is accepted by the Join operation.
  parameters:
    RequestIdHeader:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller correlation ID; the server returns the accepted/generated value in response header and meta.requestId.
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: '^[A-Za-z0-9._:-]+$'
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: Reuse for retries of the same normalized request; reuse with different content returns a conflict.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[\x21-\x7E]+$'
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor bound to caller scope, filters, and sort.
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      required: false
      description: Comma-separated endpoint whitelist; prefix a field with '-' for descending order.
      schema:
        type: string
        maxLength: 200
...
    ClassSectionFilter:
      name: classSectionId
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/OpaqueId'
...
    ClassSectionId:
      name: classSectionId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/OpaqueId' }
    EnrollmentId:
      name: enrollmentId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/OpaqueId' }
    InviteToken:
      name: inviteToken
      in: path
      required: true
      description: Opaque sensitive transport token; never log its full value.
      schema:
        type: string
        minLength: 16
        maxLength: 512
...
    ExportId:
      name: exportId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/OpaqueId' }
...
  responses:
...
    EnrollmentSuccess:
      description: Enrollment resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnrollmentEnvelope' }
    EnrollmentListSuccess:
      description: Paged enrollment resources
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnrollmentListEnvelope' }
...
    JoinSuccess:
      description: Atomic student profile, ACTIVE enrollment, and authentication session result
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
        Cache-Control: { schema: { type: string, const: no-store } }
        Pragma: { schema: { type: string, const: no-cache } }
        Referrer-Policy: { schema: { type: string, const: no-referrer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/JoinEnvelope' }
...
    ExportSuccess:
      description: Export transport status
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ExportEnvelope' }
    ExportListSuccess:
      description: Paged export transport statuses
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ExportListEnvelope' }
    ExportDownloadSuccess:
      description: Short-lived export download URL
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ExportDownloadEnvelope' }
...
    BadRequest:
      $ref: '#/components/responses/Error400'
    Unauthorized:
      $ref: '#/components/responses/Error401'
    Forbidden:
      $ref: '#/components/responses/Error403'
    NotFound:
      $ref: '#/components/responses/Error404'
    Gone:
      $ref: '#/components/responses/Error410'
...
    Conflict:
      $ref: '#/components/responses/Error409'
    UnprocessableEntity:
      $ref: '#/components/responses/Error422'
...
    ServiceUnavailable:
      $ref: '#/components/responses/Error503'
    Error400:
      description: Bad request
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error401:
      description: Unauthorized
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error403:
      description: Forbidden
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error404:
      description: Not found
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error409:
      description: Conflict
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error410:
      description: Gone
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Error422:
      description: Unprocessable entity
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Error503:
      description: Service unavailable, maintenance, or read-only mode
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
        Retry-After:
          description: Seconds until a retry may be attempted when known.
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  headers:
    RequestIdResponseHeader:
      description: Server-accepted/generated request ID; equals meta.requestId on success and requestId on error.
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
  schemas:
    OpaqueId:
      type: string
      minLength: 1
      maxLength: 64
      pattern: '^[A-Za-z0-9._:-]+$'
      description: Opaque internal ID; clients must not parse prefixes or ordering.
...
    PositiveVersion:
      type: integer
      format: int64
      minimum: 1
...
    EnrollmentStatus:
      type: string
      enum: [ACTIVE, WITHDRAWN, REMOVED]
    EnrollmentSource:
      type: string
      enum: [OFFICIAL_IMPORT, QR_CODE, MANUAL, SYSTEM_SYNC]
...
    SuccessMeta:
      type: object
      additionalProperties: false
      properties:
        requestId:
          type: string
          minLength: 1
          maxLength: 64
      required: [requestId]
    PagedMeta:
      type: object
      additionalProperties: false
      properties:
        requestId:
          type: string
          minLength: 1
          maxLength: 64
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
      required: [requestId, pagination]
    ErrorEnvelope:
      type: object
      additionalProperties: false
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message:
          type: string
          minLength: 1
          maxLength: 1000
        details: { $ref: '#/components/schemas/ErrorDetails' }
        requestId:
          type: string
          minLength: 1
          maxLength: 64
        timestamp:
          type: string
          format: date-time
      required: [code, message, details, requestId, timestamp]
      example:
        code: CONFLICT_VERSION_MISMATCH
        message: The resource changed. Refresh and try again.
        details:
          resourceType: EXERCISE_RECORD
          resourceId: rec_01JABC123
          expectedVersion: 3
          actualVersion: 4
        requestId: req_01JABC123
        timestamp: '2026-08-02T12:00:00Z'
...
    Enrollment:
      type: object
      additionalProperties: false
      properties:
        id: { $ref: '#/components/schemas/OpaqueId' }
        organizationId: { $ref: '#/components/schemas/OpaqueId' }
        semesterId: { $ref: '#/components/schemas/OpaqueId' }
        classSectionId: { $ref: '#/components/schemas/OpaqueId' }
        studentId: { $ref: '#/components/schemas/OpaqueId' }
        source: { $ref: '#/components/schemas/EnrollmentSource' }
        sourceReferenceId: { type: [string, 'null'], maxLength: 64 }
        status: { $ref: '#/components/schemas/EnrollmentStatus' }
        joinedAt: { type: string, format: date-time }
        endedAt: { type: [string, 'null'], format: date-time }
        endReason: { type: [string, 'null'], maxLength: 500 }
        createdBy: { type: [string, 'null'], maxLength: 64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { $ref: '#/components/schemas/PositiveVersion' }
      required: [id, organizationId, semesterId, classSectionId, studentId, source, sourceReferenceId, status, joinedAt, endedAt, endReason, createdBy, createdAt, updatedAt, version]
...
    ExportType:
      type: string
      enum: [ROSTER_ALIGNMENT, EXERCISE_RECORDS, STUDENT_SCORES, AUDIT_LOGS]
    ExportStatus:
      type: string
      pattern: '^[A-Z][A-Z0-9_]*$'
      description: Open contract-only value; V1 has no ExportJob lifecycle because every Export operation is fail-closed.
      x-unresolved: EXPORT_GATE
...
    VersionedReasonRequest:
      type: object
      additionalProperties: false
      required: [reason, expectedVersion]
      properties:
        reason: { type: string, minLength: 1, maxLength: 1000 }
        expectedVersion: { $ref: '#/components/schemas/PositiveVersion' }
    ManualEnrollmentRequest:
      type: object
      additionalProperties: false
      required: [studentId, reason]
      properties:
        studentId: { $ref: '#/components/schemas/OpaqueId' }
        reason: { type: string, minLength: 1, maxLength: 1000 }
...
    CreateExportRequest:
      type: object
      additionalProperties: false
      required: [exportType, filters, purpose]
      properties:
        exportType: { $ref: '#/components/schemas/ExportType' }
        filters: { type: object, additionalProperties: true }
        purpose: { type: string, minLength: 1, maxLength: 500 }
      description: Contract-only request; V1 fails closed with SYSTEM_MODE_UNSUPPORTED and does not persist an ExportJob.
...
    EnrollmentEnvelope: { $ref: '#/components/schemas/EnrollmentResponse' }
    EnrollmentListEnvelope: { $ref: '#/components/schemas/EnrollmentListResponse' }
...
    JoinEnvelope: { $ref: '#/components/schemas/JoinResponse' }
...
    ExportEnvelope: { $ref: '#/components/schemas/ExportResponse' }
    ExportListEnvelope: { $ref: '#/components/schemas/ExportListResponse' }
    ExportDownloadEnvelope: { $ref: '#/components/schemas/ExportDownloadResponse' }
...
    EnrollmentResponse:
      type: object
      required: [data, meta]
      properties: { data: { $ref: '#/components/schemas/Enrollment' }, meta: { $ref: '#/components/schemas/SuccessMeta' } }
    EnrollmentListResponse:
      type: object
      required: [data, meta]
      properties: { data: { type: array, items: { $ref: '#/components/schemas/Enrollment' } }, meta: { $ref: '#/components/schemas/PagedMeta' } }
...
    JoinResponse:
      type: object
      required: [data, meta]
      properties: { data: { $ref: '#/components/schemas/JoinResult' }, meta: { $ref: '#/components/schemas/SuccessMeta' } }
...
    ExportResponse:
      type: object
      required: [data, meta]
      properties: { data: { $ref: '#/components/schemas/ExportJob' }, meta: { $ref: '#/components/schemas/SuccessMeta' } }
    ExportListResponse:
      type: object
      required: [data, meta]
      properties: { data: { type: array, items: { $ref: '#/components/schemas/ExportJob' } }, meta: { $ref: '#/components/schemas/PagedMeta' } }
    ExportDownloadResponse:
      type: object
      required: [data, meta]
      properties: { data: { $ref: '#/components/schemas/ExportDownload' }, meta: { $ref: '#/components/schemas/SuccessMeta' } }
...